pyproject.toml
[project]
name = "testproproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "isort>=6.0.1",
    "ruff>=0.11.7",
]
